libcopybit_wrapper: only request channel once.
-although the function name here is fill_color this does not
match the name of the function inside the real copybit lib.
The name of the function inside the real copybit lib is something
like hwc_request_chanell_open and there is an equivalent
hwc_request_chanell_close which we are blocking to prevent
a fatal crash. Since we are not keeping an equalibrium between
opens and closes we must also not allow a double open.

Signed-off-by: Thespartann <danilastefan69@yahoo.ro>